Added a new java class to BF-CPP - BioFormatsExtensionPrinter - to generate a text file containing a list of all file extensions supported by Bio-Formats.
Modified the ant scripts for BF-CPP and BF-ITK to automatically
compile and run this class, and to move the resulting text file
to the bf-itk/dist location with the rest of the BF-ITK plugin.

Also, added a "clean-all" target to BF-CPP which cleans the
normal BF-CPP build and the proxies.