Returned loci common to branch
The moving of loci common source files to scifio rendered this branch
unusable.  For the duration of development these classes have been
copied back to loci common (they still exist, with an ome.scifio package
prefix, in the scifio component).

As the current goal is to establish the scifio api, the final logistics
of build structure will be deferred until the api is complete.

Note that the former loci common classes that remain on the scifio
branch were updated to depend on the scifio versions of their prior
dependencies.