Updated API with implementation of APNGChecker
In this first stage of converting an existing Bio-Formats reader (in
this case, for the APNG format) the checker was added.

Note that the API for all classes is being adjusted through this process
as decisions are made as to where to put original Bio-Formats
functionality in the new SCIFIO framework.