Restored computeCRC method
The computeCRC method was mistakenly removed from APNG reader when
refactoring the PNG block subclass (moved to its own class in
ome.scifio.in.apng).  It is now returned.

Note that code that is currently non-functional in APNGReader has been
commented out (as the class is in transition to SCIFIO-component
representation).