Added first pass at APNGReader
With this commit,t he APNGReader is functional within SCIFIO.  Output
for an openBytes call is identical to the original loci.formats reader.

The original loci.formats.in.APNGReader now defers functionality to
ome.scifio.apng.APNGReader.

However: note that some functionality may be absent from the APNGReader
until the abstract superclasses are ultimately converted to SCIFIO
format, due to the separatation of variables (e.g. openBytes works, but
calling reader.isInterleaved() probably will not because the
loci.formats core doesn't have actual image data to read, as it is
stored on the SCIFIO side... but as isInterleaved() is an abstract
superclass method, it does not currently defer to SCIFIO).