SCIFIO Metadata -> resetMeta() functionality
Added a resetMeta() method to recursively delete all data members of a
Metadata object and return it to a state as if just instantiated.
resetMeta() recursively descends through the related superclasses and
interfaces, resetting all data members.

The implementation lives in ome.scifio.AbstractMetadata.java.  Interface
and subclasses were updated as well to take advantage of this new
functionality.

Fundamentally, this was precipitated by a desire to to refactor the
Translator component to accept two Metadata objects to allow Translator
chaining.