loci.formats.APNGWriter defers to SCIFIO
Removed functional code from loci.formats.out.APNGWriter, instead deferring
to ome.scifio.out.apng.APNGWriter.

Note: As FormatWriter and FormatReader have not been converted to SCIFIO
deference, there are methods in loci.formats.out.APNGWriter that can and
should be removed when FormatWriter and the rest of the
Writers are converted to SCIFIO components.