ome.scifio.Writer accepts image index for init
In loci.formats.FormatReader, the current series was used to initialize
Writers (e.g. for APNGWriter).

With removal of state in SCIFIO, an image index can be supplied to be
used as the basis for initialzation.  If none is specified, 0 is used.

Note: initialization in SCIFIO is similar to setId in classic
Bio-Formats.