Added getters and setters for SCIFIO Metadata
Added the full compliment of getters and setters in the
ome.scifio.Metadata API. Corresponding implementations were added for
current children of Metadata.

These (now mandatory, API defined) getters and setters represent the
minimal fields capturd in metadata objects.

The primary motivation for adding these methods is to preserve
functionality in the SCIFIOFormatReader and SCIFIOFormatWriter classes.