Implemented SCIFIO-deferring readers/writers
Added complete implmenetations for loci.formats SCIFIOFormatReader,
SCIFIOFormatWriter, and loci.formats.in SCIFIOBIFormatReader

These classes are all deprecated and serve as an intermediary between
the original FormatReader and FormatWriter, and implementations of
specific Reader classes that are deferring to SCIFIO.

This allows converted Readers to defer all functionality to SCIFIO
(by effectively breaking the dependency to the abstract implementations
of FormatReader and FormatWriter and passing control to the SCIFIO
abstract layer). Un-converted Readers, however, retain their
functionality by continuing to use the FormatReader and FormatWriter
implementations.

When all loci.formats Readers and Writers have been converted to SCIFIO
components, the original FormatReader and FormatWriter level can be
removed so only the SCIFIO versions remain.