Mirrored MetadataOptions functionality in SCIFIO
refs #7702

Added versions of DefaultMetadataOptions, MetadataLevel and
MetadataOptions to ome.scifio. Going forward, these will be the classes
usd to describe the metadata collection options.

The loci.formats.in version of these classes were modified to convert
and defer to the ome.scifio versions, while returning appropriately
typed data structures to preserve core functionality.

This conversion was necessary to preserve functionality of the
SCIFIOFormatReader class.